requirements.txt
Django==1.9.7
Markdown==2.6.6
django-filter==0.13.0
djangorestframework==3.4.0
pytz==2016.6.1
